[TEST] Extracts duplicated test helpers into a shared scope.
