+++ EssentialFeedTests/Helpers/SharedTestsHelpers.swift
@@ -1,0 +1,13 @@
+//
+//  Copyright © 2020 Jesús Alfredo Hernández Alarcón. All rights reserved.
+//
+
+import Foundation
+
+func anyNSError() -> NSError {
+    NSError(domain: "any error", code: 1)
+}
+
+func anyURL() -> URL {
+    URL(string: "http://any-url.com")!
+}
